server/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
mccabe==0.6.1
msgpack==1.0.3
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
pytest-asyncio==0.16.0
pytest-django==4.5.2
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.1.0
txaio==21.2.1
typing-extensions==4.1.0
zope.interface==5.4.0
